chore: extract config watcher and rate limiter
